Move to extra querygraph package
