+++ core/src/main/java/com/graphhopper/routing/querygraph/QueryGraphTurnExt.java
@@ -16,7 +16,7 @@
  *  limitations under the License.
  */
 
-package com.graphhopper.routing;
+package com.graphhopper.routing.querygraph;
 
 import com.carrotsearch.hppc.IntArrayList;
 import com.graphhopper.storage.Graph;
