+++ core/src/test/java/com/graphhopper/storage/GraphHopperStorageCHTest.java
@@ -17,9 +17,9 @@
  */
 package com.graphhopper.storage;
 
-import com.graphhopper.routing.QueryGraph;
 import com.graphhopper.routing.ch.PrepareEncoder;
 import com.graphhopper.routing.profiles.BooleanEncodedValue;
+import com.graphhopper.routing.querygraph.QueryGraph;
 import com.graphhopper.routing.util.*;
 import com.graphhopper.routing.weighting.FastestWeighting;
 import com.graphhopper.routing.weighting.ShortestWeighting;
